Added a little documentation about the config.status file and what it does.
Reworked the wrapper commands: when the configure script tries to work
out which compiler is being used to build openshmem, it then passes
along the corresponding C++ and Fortran commands, which get substituted
in the wrapper scripts.  Fixes problem with the Fortran wrapper and
Intel compiler.  Detection can still be improved.

TODO in Makefile.in notes that there's scope for improvement since most
of the work is repeated for each wrapper.