correcting check for GASNET_RCV_THREAD
Should only check for GASNET_RCV_THREAD if using the ibv or vapi GASNet
conduit. Also, if GASNET_RCV_THREAD is set to 0 or isn't set at all, use the
initial value of handle_own_thread instead of setting handle_own_thread to 0.
This makes it easier to disable the progress threads (simply initialize
handle_own_thread to 0 instead of 1).