Fixes for static linking
In Makefile.in, move the libelf -L flags over to LDFLAGS.

In oshcc.in, the GASNet LDFLAGS need to be specified before the
OpenSHMEM LDFLAGS so that library paths are searched in the correct
order.

In configure, don't ever set -Wl,-rpath.  This overrides the user's
LD_LIBRARY_PATH, which can cause problems when using local installations
of libraries that are also installed on the base system.  Specifically,
this forced the system PSM installed in /usr/lib64 to be used instead of
a local PSM installation under $HOME even though it was specified in
LD_LIBRARY_PATH.

These changes correct the linker argument order, which is now more
strict due to linking OpenSHMEM statically (and specifying all 3rd party
dependencies on the application's link command).  GASNet library search
paths (LDFLAGS) are specified first, followed by OpenSHMEM's.  The
user's command line arguments are then inserted.  Finally, libraries are
specified in the reverse order from the LDFLAGS.  OpenSHMEM first, then
GASNet.

Signed-off-by: Andrew Friedley <andrew.friedley@intel.com>