Make progress thread configurable
Add a configure script option to enable/disable the OpenSHMEM progress
thread.  The progress thread is enabled by default, as it was before.

GASNet-PSM already has (more desirable) progress threads at the PSM and
GASNet layers, so there is no need for an OpenSHMEM progress thread.
This patch allows OpenSHMEM to be configured for better performance in
this case.

Signed-off-by: Andrew Friedley <andrew.friedley@intel.com>