Revert to linear barrier, zap buggy tree one
Deepak says:

I have a code that uses lots of reductions and it runs correctly using
Open MPI SHMEM, but it gives inconsistent results with our implementation
when the barrier algorithm is set to tree (the default). When I switch
to the linear barrier, the problem goes away.

So, inspecting the barrier-tree algorithm, what happens is every PE in
the active set occupies a node in a binary tree. For 8 PEs, all in the
active set, it looks like:

0
|   \
1    2 __
| \   \   \
3  4   5   6
|
7

The way the algorithm works, notifications will be first sent down the
tree, starting from the root, using pSync[0] (before a PE sends out the
notification, it waits for a notification from its parent and then also
waits for each of its children to enter the barrier).

Once these downstream notifications have completed, notifications are
sent up to the parent starting at the leaves using pSync[1].  So, each
PE will wait for notifications from all its child nodes (at most 2),
and then send a notification up to its parent.

This algorithm isn't correct for the above tree, however. Consider PE
3. It only ends up waiting for 0, 1, and 7.

The attached test code illustrates the problem when run with 8 PEs. You
can see that PEs 7, 3, 1, and 5 all pass through the barrier without
waiting for PE 6, which is stuck in a sleep call for 20 seconds prior
to the barrier.

Besides this error, the implementation also has the PEs spinning on
remote pSync's waiting for the children to arrive at the barrier. It
seems to me that it was translated from a shared memory implementation
where such a thing may be OK. But this is a bad choice for our case,
when the remote pSync is likely over the network.

Since this is a serious error, we should probably change the default
barrier implementation to linear until barrier-tree is fixed (or we come
up with another barrier algorithm like dissemination which may be more
suitable for distributed memory).